+++ interfaces/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0 - 03 June 2026
+
+* Add capability to encrypt and decrypt specific fields and store them in relation data
+* In cross-model relations, store 'mtls-cert' encrypted to relation data instead of a secret
+* requires `cryptography` >= 48.0.0 to be added to a charms dependencies
+
 # 1.0.2 - 31 March 2026
 
 * Fix import for `ValkeyResponseModel`